Fixed: missing translation of fishing skill (introducted in Loki with Hildir but never translated, resulting in fishing being shown as Unknown).
Note: not using existing translation for Fish since it refers to the item fish rather than the skill to fish. It works in English to use the same but it might not do so in other languages. Silly detail, I know (since it likely will not be translated to more languages ever).